Extract some classes into their own files
