+++ Hudl.Mjolnir.Tests/Breaker/FailurePercentageCircuitBreakerTests.cs
@@ -1,7 +1,9 @@
 ﻿using Hudl.Mjolnir.Breaker;
 using Hudl.Mjolnir.Clock;
+using Hudl.Mjolnir.Events;
 using Hudl.Mjolnir.External;
 using Hudl.Mjolnir.Key;
+using Hudl.Mjolnir.Log;
 using Hudl.Mjolnir.Metrics;
 using Hudl.Mjolnir.Tests.Helper;
 using Moq;
