+++ Hudl.Mjolnir.Tests/Metrics/StandardCommandMetricsTests.cs
@@ -1,6 +1,6 @@
 ﻿using Hudl.Mjolnir.Clock;
-using Hudl.Mjolnir.External;
 using Hudl.Mjolnir.Key;
+using Hudl.Mjolnir.Log;
 using Hudl.Mjolnir.Metrics;
 using Hudl.Mjolnir.Tests.Helper;
 using Moq;
