+++ Hudl.Mjolnir/Hudl.Mjolnir.csproj
@@ -50,6 +50,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Breaker\FailurePercentageCircuitBreakerConfig.cs" />
+    <Compile Include="Bulkhead\BulkheadConfig.cs" />
     <Compile Include="Bulkhead\BulkheadRejectedException.cs" />
     <Compile Include="Bulkhead\IBulkheadSemaphore.cs" />
     <Compile Include="Bulkhead\SemaphoreBulkhead.cs" />
@@ -67,11 +69,14 @@
     <Compile Include="Command\CommandInvoker.cs" />
     <Compile Include="Command\CommandResult.cs" />
     <Compile Include="Command\SyncCommand.cs" />
-    <Compile Include="External\IConfig.cs" />
-    <Compile Include="External\IgnoringMetricEvents.cs" />
+    <Compile Include="Config\DefaultValueConfig.cs" />
+    <Compile Include="External\IMjolnirConfig.cs" />
+    <Compile Include="Events\IgnoringMetricEvents.cs" />
     <Compile Include="External\IMjolnirLog.cs" />
     <Compile Include="External\IMetricEvents.cs" />
     <Compile Include="Key\GroupKey.cs" />
+    <Compile Include="Log\DefaultMjolnirLog.cs" />
+    <Compile Include="Log\DefaultMjolnirLogFactory.cs" />
     <Compile Include="Metrics\CounterMetric.cs" />
     <Compile Include="Metrics\ICommandMetrics.cs" />
     <Compile Include="Metrics\ILongCounter.cs" />
@@ -81,6 +86,7 @@
     <Compile Include="Metrics\ResettingNumbersBucket.cs" />
     <Compile Include="Metrics\StandardCommandMetrics.cs" />
     <Compile Include="Command\CommandCompletionStatus.cs" />
+    <Compile Include="Metrics\StandardCommandMetricsConfig.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Util\ConcurrentDictionaryExtensions.cs" />
     <Compile Include="Util\GaugeTimer.cs" />
