+++ Hudl.Mjolnir/Log/DefaultMjolnirLog.cs
@@ -1,0 +1,25 @@
+﻿using Hudl.Mjolnir.External;
+using System;
+
+namespace Hudl.Mjolnir.Log
+{
+    // TODO is ignoring the right default? maybe a console log instead?
+
+    internal class DefaultMjolnirLog : IMjolnirLog
+    {
+        public void Error(string message)
+        {
+            return;
+        }
+
+        public void Error(string message, Exception exception)
+        {
+            return;
+        }
+
+        public void Info(string message)
+        {
+            return;
+        }
+    }
+}
